apply the second review round
Findings must name a location: no rule produces one without it, so the
empty-location branch of __str__ guarded a case that cannot occur — the
same standard the earlier try/except was held to.

The object enumeration takes the type to look for, iter_objects(database,
of_type), so further rules can reuse it for other kinds of objects;
iter_codecs remains as the typed convenience both current rules use.

The rule classes are no longer part of the package's public names: they
are implementation details of DEFAULT_RULES and remain importable from
their own modules.
