+++ odxtools/checks/__init__.py
@@ -21,8 +21,6 @@ from .rule import Rule
 __all__ = [
     "DEFAULT_RULES",
     "Finding",
-    "NrcConstWithoutValue",
-    "OverlappingParameters",
     "Rule",
     "Severity",
     "run_checks",
